+++ src/aws_organization/lib/__init__.py
@@ -1,4 +1,5 @@
 from .account import AwsAccount
+from .central_infra_workload import create_central_infra_workload
 from .org_units import OrganizationalUnits
 from .org_units import create_organizational_units
 from .workload import DEFAULT_ORG_ACCESS_ROLE_NAME
